performance optimization for service top-n (#2363)
* feat(msp): performance optimization for service top-n

* feat(msp): remove old service view

feat(msp): remove old service view

(cherry picked from commit d44f832212cf7d68cbd593d22c5e5bee9636b675)
