+++ shell/app/common/components/empty-holder/index.tsx
@@ -40,7 +40,7 @@ export const EmptyHolder = ({
   });
   return (
     <div className={`${cls} ${className}`} style={style}>
-      {typeof icon === 'string' ? <CustomIcon type={icon} color /> : <div>{icon}</div>}
+      {typeof icon === 'string' ? <CustomIcon className="w-full" type={icon} color /> : <div>{icon}</div>}
       <span>
         {tip} <span className="action">{action}</span>
       </span>
