+++ shell/app/config-page/components/top-n/index.scss
@@ -1,8 +1,13 @@
 .cp-data-rank {
+  .items-wrapper {
+    min-height: 232px;
+  }
+
   .bg-icon-wrapper {
     width: 52px;
     height: 52px;
   }
+
   .ant-progress,
   .ant-progress-outer {
     line-height: 4px;
